fix: eliminate warnings in example and test code
- Fix conversion warnings in example code (int to uint16_t)
- Fix unused parameter warnings in test lambda functions
- Fix conversion warnings in test performance calculations
- Fix conversion warnings in integration tests

Examples and core tests now build with significantly fewer warnings.
Core library remains completely warning-free.

+++ examples/tutorials/tcp_server/echo_server.cpp
@@ -120,7 +120,7 @@ int main(int argc, char** argv) {
   std::signal(SIGTERM, signal_handler);
 
   // Parse port
-  uint16_t port = (argc > 1) ? std::stoi(argv[1]) : 8080;
+  uint16_t port = (argc > 1) ? static_cast<uint16_t>(std::stoi(argv[1])) : 8080;
 
   // Create and start server
   EchoServer server(port);
