+++ test/unit/common/test_memory.cc
@@ -160,7 +160,7 @@ TEST_F(MemoryIntegratedTest, MemoryPoolPerformance) {
   auto end_time = std::chrono::high_resolution_clock::now();
   auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end_time - start_time);
 
-  double throughput = static_cast<double>(num_operations) / (duration.count() / 1000000.0);
+  double throughput = static_cast<double>(num_operations) / (static_cast<double>(duration.count()) / 1000000.0);
 
   std::cout << "Memory pool performance:" << std::endl;
   std::cout << "  Operations: " << num_operations << std::endl;
